+++ src/grouplens.hpp
@@ -13,7 +13,7 @@ float grouplens(const V &avg_product_rating, // avg_product_rating[i] - average 
 	float numer = 0;
 	float denom = 0;
 	
-	for (size_t i=0; i<users_rating.rows(); ++i)
+	for (int i=0; i<users_rating.rows(); ++i)
 	{
 		float user_resemblance = resemblance(user,i);
 		
@@ -29,9 +29,9 @@ void grouplens(M &grouplens_predict, const M &users_ratings,
 				const M &user_resemblance, 
 				const V &avg_users_rating, const V &avg_product_ratings)
 {
-	for (size_t i=0; i<users_ratings.rows(); ++i)	// users
+	for (int i=0; i<users_ratings.rows(); ++i)	// users
 	{
-		for (size_t j=0; j<users_ratings.cols(); ++j)	// products
+		for (int j=0; j<users_ratings.cols(); ++j)	// products
 		{
 			grouplens_predict(i,j) = grouplens(avg_product_ratings, 
 										users_ratings,
